requirements.txt
attrs==21.4.0
bleach==6.1.0
bokeh==3.4.2
boto3==1.34.137
botocore==1.34.137
cachetools==5.0.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==7.1.2
contourpy==1.2.1
cryptography==36.0.2
Flask==1.1.4
idna==3.7
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==1.0.1
linkify-it-py==2.0.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdit-py-plugins==0.4.1
mdurl==0.1.2
numpy==2.0.0
packaging==24.1
pandas==2.2.2
panel==1.4.4
param==2.1.1
pillow==10.4.0
pycparser==2.22
PyJWT==2.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyviz_comms==3.0.2
PyYAML==6.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.2
webencodings==0.5.1
Werkzeug==1.0.1
xyzservices==2024.6.0
